batch: Move realized range mapping into package
The realization package owns entry data, provenance, storage, and
boundaries while working-range construction remains at the batch package
root.

Working-range construction writes directly into realized storage using line
mappings, making the flat location an incomplete account of the
representation.

This commit moves realized range mapping beneath realization and points it
directly at nested matching and entry-storage modules.

A tests-only follow-up will adopt the realization imports before later
batch packages consume them.
